:recycle: Add utils module with findGitmojiCommand
